+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to project will be documented in this file.
 
+## [0.10.5] - 2026-04-19
+### Fixed
+- Model structure view not updated when opening the project
+- Extra transition created when changing transition source
+- Delete linked transitions when deleting a state
+- Transitions incorrectly rendered after resizing parent state with substate movement
+
 ## [0.10.4] - 2026-04-18
 ### Fixed
 - self-transition for a substate is added twice